+++ include/h.h
@@ -968,6 +968,7 @@ extern MODVAR int (*get_connections_from_ip)(Client *client);
 extern MODVAR int (*get_floodprot_channel_max_lines)(Channel *channel);
 extern MODVAR int (*floodprot_check_multiline_batch)(Channel *channel, Client *client, int line_count);
 extern MODVAR int (*channel_flood_blocked_count)(Client *client, const char *type);
+extern MODVAR void (*channel_flood_expand_json)(json_t *root, Client *client);
 /* /Efuncs */
 
 /* TLS functions */
@@ -1010,6 +1011,7 @@ extern int get_connections_from_ip_default_handler(Client *client);
 extern int get_floodprot_channel_max_lines_default_handler(Channel *channel);
 extern int floodprot_check_multiline_batch_default_handler(Channel *channel, Client *client, int line_count);
 extern int channel_flood_blocked_count_default_handler(Client *client, const char *type);
+extern void channel_flood_expand_json_default_handler(json_t *root, Client *client);
 extern void do_unreal_log_remote_deliver_default_handler(LogLevel loglevel, const char *subsystem, const char *event_id, MultiLine *msg, const char *json_serialized);
 extern int make_oper_default_handler(Client *client, const char *operblock_name, const char *operclass, ConfigItem_class *clientclass, long modes, const char *snomask, const char *vhost, const char *autojoin_channels);
 extern void webserver_send_response_default_handler(Client *client, int status, char *msg);
@@ -1342,6 +1344,7 @@ extern void json_expand_client_security_groups(json_t *parent, Client *client);
 extern void json_expand_channel(json_t *j, const char *key, Channel *channel, int detail);
 extern void json_expand_tkl(json_t *j, const char *key, TKL *tkl, int detail);
 extern void json_expand_textanalysis(json_t *root, const char *key, TextAnalysis *ta, int detail);
+extern void json_expand_flood_counts(json_t *root, const char *key, Client *client);
 extern void json_expand_mask_list(json_t *parent, const char *key, ConfigItem_mask *mask);
 extern void json_expand_name_list(json_t *parent, const char *key, NameList *list);
 extern void json_expand_nvplist(json_t *parent, const char *key, NameValuePrioList *list);