+++ src/api-efunctions.c
@@ -199,6 +199,7 @@ int (*get_connections_from_ip)(Client *client);
 int (*get_floodprot_channel_max_lines)(Channel *channel);
 int (*floodprot_check_multiline_batch)(Channel *channel, Client *client, int line_count);
 int (*channel_flood_blocked_count)(Client *client, const char *type);
+void (*channel_flood_expand_json)(json_t *root, Client *client);
 
 Efunction *EfunctionAddMain(Module *module, EfunctionType eftype, int (*func)(), void (*vfunc)(), void *(*pvfunc)(), char *(*stringfunc)(), const char *(*conststringfunc)())
 {
@@ -560,4 +561,5 @@ void efunctions_init(void)
 	efunc_init_function(EFUNC_GET_FLOODPROT_CHANNEL_MAX_LINES, get_floodprot_channel_max_lines, get_floodprot_channel_max_lines_default_handler, 0);
 	efunc_init_function(EFUNC_FLOODPROT_CHECK_MULTILINE_BATCH, floodprot_check_multiline_batch, floodprot_check_multiline_batch_default_handler, 0);
 	efunc_init_function(EFUNC_CHANNEL_FLOOD_BLOCKED_COUNT, channel_flood_blocked_count, channel_flood_blocked_count_default_handler, 0);
+	efunc_init_function(EFUNC_CHANNEL_FLOOD_EXPAND_JSON, channel_flood_expand_json, channel_flood_expand_json_default_handler, 0);
 }